6 modules compile. Sample misses Henson generated classes
