+++ dart-annotations/pom.xml
@@ -11,8 +11,8 @@
     <version>2.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>dart</artifactId>
-  <name>Dart</name>
+  <artifactId>dart-annotations</artifactId>
+  <name>Dart Annotations</name>
 
   <dependencies>
     <dependency>
@@ -39,13 +39,6 @@
       <groupId>org.parceler</groupId>
       <artifactId>parceler-api</artifactId>
       <scope>test</scope>
-    </dependency>
-
-
-    <dependency>
-      <groupId>com.f2prateek.dart</groupId>
-      <artifactId>dart-annotations</artifactId>
-      <version>2.0.0-SNAPSHOT</version>
     </dependency>
 
     <dependency>
