+++ dart-common/pom.xml
@@ -8,7 +8,7 @@
   <parent>
     <groupId>com.f2prateek.dart</groupId>
     <artifactId>dart-parent</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>dart-common</artifactId>
@@ -41,6 +41,11 @@
       <scope>test</scope>
     </dependency>
 
+    <dependency>
+      <groupId>com.f2prateek.dart</groupId>
+      <artifactId>dart-annotations</artifactId>
+      <version>2.0.0-SNAPSHOT</version>
+    </dependency>
 
     <dependency>
         <groupId>com.squareup</groupId>
