+++ dart-common/src/main/java/com/f2prateek/dart/FieldBinding.java
@@ -1,0 +1,54 @@
+/*
+ * Copyright 2013 Jake Wharton
+ * Copyright 2014 Prateek Srivastava (@f2prateek)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.f2prateek.dart;
+
+import javax.lang.model.type.TypeMirror;
+
+public final class FieldBinding implements Binding {
+  private final String name;
+  private final TypeMirror type;
+  private final boolean required;
+  private final boolean parcel;
+
+  FieldBinding(String name, TypeMirror type, boolean required, boolean parcel) {
+    this.name = name;
+    this.type = type;
+    this.required = required;
+    this.parcel = parcel;
+  }
+
+  public String getName() {
+    return name;
+  }
+
+  public TypeMirror getType() {
+    return type;
+  }
+
+  @Override public String getDescription() {
+    return "field '" + name + "'";
+  }
+
+  @Override public boolean isRequired() {
+    return required;
+  }
+
+  public boolean isParcel() {
+    return parcel;
+  }
+}
