+++ dart-processor/pom.xml
@@ -11,8 +11,8 @@
     <version>2.0.0-SNAPSHOT</version>
   </parent>
 
-  <artifactId>dart</artifactId>
-  <name>Dart</name>
+  <artifactId>dart-processor</artifactId>
+  <name>Dart Processor</name>
 
   <dependencies>
     <dependency>
@@ -44,7 +44,13 @@
 
     <dependency>
       <groupId>com.f2prateek.dart</groupId>
-      <artifactId>dart-annotations</artifactId>
+      <artifactId>dart-common</artifactId>
+      <version>2.0.0-SNAPSHOT</version>
+    </dependency>
+
+    <dependency>
+      <groupId>com.f2prateek.dart</groupId>
+      <artifactId>dart</artifactId>
       <version>2.0.0-SNAPSHOT</version>
     </dependency>
