+++ dart-processor/src/test/java/com/f2prateek/dart/ProcessorTestUtilities.java
@@ -15,26 +15,14 @@
  * limitations under the License.
  */
 
-package com.f2prateek.dart.example;
+package com.f2prateek.dart;
 
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Bundle;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import com.f2prateek.dart.Dart;
+import com.f2prateek.dart.AbstractDartProcessor;
+import java.util.Arrays;
+import javax.annotation.processing.Processor;
 
-public class MainActivity extends Activity {
-
-  @Override protected void onCreate(Bundle savedInstanceState) {
-    super.onCreate(savedInstanceState);
-
-    setContentView(R.layout.activity_main);
-
-    ButterKnife.inject(this);
-    Dart.inject(this);
-  }
-
-  @OnClick(R.id.button) public void onLaunchButtonClick() {
+final class ProcessorTestUtilities {
+  static Iterable<? extends Processor> dartProcessors() {
+    return Arrays.asList(new InjectExtraProcessor());
   }
 }
