+++ dart-sample/pom.xml
@@ -8,7 +8,7 @@
   <parent>
     <groupId>com.f2prateek.dart</groupId>
     <artifactId>dart-parent</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>dart-sample</artifactId>
@@ -18,7 +18,22 @@
   <dependencies>
     <dependency>
       <groupId>com.f2prateek.dart</groupId>
+      <artifactId>dart-processor</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.f2prateek.dart</groupId>
       <artifactId>dart</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.f2prateek.dart</groupId>
+      <artifactId>henson-processor</artifactId>
+      <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.f2prateek.dart</groupId>
+      <artifactId>henson</artifactId>
       <version>${project.version}</version>
     </dependency>
