+++ henson-processor/src/main/java/com/f2prateek/dart/henson/HensonExtraProcessor.java
@@ -1,0 +1,68 @@
+/*
+ * Copyright 2013 Jake Wharton
+ * Copyright 2014 Prateek Srivastava (@f2prateek)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *        http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.f2prateek.dart.henson;
+
+import com.f2prateek.dart.AbstractDartProcessor;
+import com.f2prateek.dart.InjectionTarget;
+import java.io.IOException;
+import java.io.Writer;
+import java.util.Map;
+import java.util.Set;
+import javax.annotation.processing.RoundEnvironment;
+import javax.lang.model.element.TypeElement;
+import javax.tools.JavaFileObject;
+
+public final class HensonExtraProcessor extends AbstractDartProcessor {
+  @Override public boolean process(Set<? extends TypeElement> elements, RoundEnvironment env) {
+    Map<TypeElement, InjectionTarget> targetClassMap = findAndParseTargets(env);
+
+    System.out.println("starting henson");
+    for (Map.Entry<TypeElement, InjectionTarget> entry : targetClassMap.entrySet()) {
+      TypeElement typeElement = entry.getKey();
+      InjectionTarget injectionTarget = entry.getValue();
+
+      System.out.println("henson processes " + injectionTarget.className);
+      // Now write the IntentBuilder
+      Writer writer = null;
+
+      // Generate the IntentBuilder
+      try {
+        BundlerBuilder intentBuilder = new BundlerBuilder(injectionTarget);
+        JavaFileObject jfo = filer.createSourceFile(intentBuilder.getFqcn(), typeElement);
+        writer = jfo.openWriter();
+        System.out.println("henson creates " + intentBuilder.brewJava() + "in " + intentBuilder.getFqcn());
+        writer.write(intentBuilder.brewJava());
+      } catch (IOException e) {
+        error(typeElement, "Unable to write intent builder for type %s: %s", typeElement,
+              e.getMessage());
+      } finally {
+        if (writer != null) {
+          try {
+            writer.close();
+          } catch (IOException e) {
+            error(typeElement, "Unable to close intent builder source file for type %s: %s",
+                  typeElement,
+                  e.getMessage());
+          }
+        }
+      }
+    }
+
+    return true;
+  }
+}
