+++ henson/pom.xml
@@ -8,7 +8,7 @@
   <parent>
     <groupId>com.f2prateek.dart</groupId>
     <artifactId>dart-parent</artifactId>
-    <version>1.1.1-SNAPSHOT</version>
+    <version>2.0.0-SNAPSHOT</version>
   </parent>
 
   <artifactId>henson</artifactId>
@@ -45,7 +45,7 @@
     <dependency>
       <groupId>com.f2prateek.dart</groupId>
       <artifactId>dart-common</artifactId>
-      <version>1.1.1-SNAPSHOT</version>
+      <version>2.0.0-SNAPSHOT</version>
     </dependency>
 
     <dependency>
